+++ sqlparser/lib/src/analysis/schema/result_set.dart
